+++ arbor/client.py
@@ -16,7 +16,7 @@ from contextlib import closing
 from typing import Any, Optional
 
 from arbor.cli import start_server, stop_server
-from arbor.server.utils.logging import get_logger
+from arbor.core.logging import get_logger
 
 logger = get_logger(__name__)
