+++ arbor/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core configuration and logging for Arbor."""
