+++ arbor/core/logging.py
@@ -36,9 +36,8 @@ class ArborFormatter(logging.Formatter):
         "arbor.server.services.managers.health_manager": "health",
         "arbor.server.services.managers.job_manager": "jobs",
         "arbor.server.services.managers.file_train_manager": "train",
-        "arbor.server.services.scripts.sft_training": "sft",
-        "arbor.server.services.scripts.grpo_training": "grpo",
-        "arbor.server.services.scripts.mmgrpo_training": "mmgrpo",
+        "arbor.training.sft.trainer": "sft",
+        "arbor.training.grpo.trainer": "grpo",
         "arbor.server.services.inference.vllm_client": "vllm",
         "arbor.server.services.inference.vllm_serve": "vllm-srv",
         "arbor.server.services.jobs.inference_job": "inf-job",
