+++ arbor/integrations/__init__.py
@@ -1,0 +1,5 @@
+"""Third-party integrations for Arbor.
+
+This package groups optional provider integrations under clear namespaces
+to keep the top-level API tidy (e.g., `arbor.integrations.dspy`).
+"""
