+++ arbor/integrations/dspy/__init__.py
@@ -1,4 +1,7 @@
-"""DSPy-compatible helpers exposed by Arbor."""
+"""DSPy integration for Arbor.
+
+Exports Arbor's DSPy provider and GRPO trainer in a structured namespace.
+"""
 
 from .arbor_provider import ArborProvider
 from .grpo_optimizer import ArborGRPO
