+++ arbor/server/api/routes/files.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 from fastapi import APIRouter, Body, File, HTTPException, Request, UploadFile
 
-from arbor.server.api.models.schemas import FileModel, PaginatedResponse
+from arbor.server.api.schemas import FileModel, PaginatedResponse
 from arbor.server.services.managers.file_manager import FileManager, FileValidationError
 
 # https://platform.openai.com/docs/api-reference/files/list
