+++ arbor/server/api/routes/grpo.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Request
 
-from arbor.server.api.models.schemas import (
+from arbor.server.api.schemas import (
     GRPOBaseRequest,  # TODO: These should be handled with a subclass like GRPOTerminateRequest
     GRPOCheckpointRequest,
     GRPOInitializeRequest,
