+++ arbor/server/api/routes/inference.py
@@ -1,12 +1,12 @@
 from fastapi import APIRouter, Request
 
-from arbor.server.api.models.schemas import (
+from arbor.core.logging import get_logger
+from arbor.server.api.schemas import (
     InferenceLaunchRequest,
     InferenceTerminateRequest,
 )
 from arbor.server.services.managers.inference_manager import InferenceManager
-from arbor.server.utils.helpers import strip_prefix
-from arbor.server.utils.logging import get_logger
+from arbor.utils.helpers import strip_prefix
 
 logger = get_logger(__name__)
