+++ arbor/server/services/jobs/inference_job.py
@@ -8,14 +8,14 @@ from typing import Optional
 import psutil
 import requests
 
-from arbor.server.core.config import Config
+from arbor.core.config import Config
+from arbor.core.logging import get_logger
 from arbor.server.services.comms.control_server import TrainerControlServer
 from arbor.server.services.inference.vllm_client import VLLMClient
 from arbor.server.services.jobs.inference_launch_config import InferenceLaunchConfig
 from arbor.server.services.jobs.job import Job, JobArtifact
-from arbor.server.utils.helpers import get_free_port
-from arbor.server.utils.logging import get_logger
-from arbor.server.utils.process_runner import InferenceProcessRunner
+from arbor.utils.helpers import get_free_port
+from arbor.utils.process_runner import InferenceProcessRunner
 
 logger = get_logger(__name__)
