+++ arbor/server/services/jobs/job.py
@@ -7,9 +7,9 @@ from typing import Callable, Literal, Optional
 
 import coolname
 
-from arbor.server.api.models.schemas import JobStatus, JobStatusModel
-from arbor.server.core.config import Config
-from arbor.server.utils.logging import get_logger
+from arbor.core.config import Config
+from arbor.core.logging import get_logger
+from arbor.server.api.schemas import JobStatus, JobStatusModel
 
 logger = get_logger(__name__)
