+++ arbor/training/__init__.py
@@ -1,0 +1,1 @@
+"""Training packages and utilities for Arbor (GRPO, DPO, SFT)."""
