+++ arbor/training/grpo/__init__.py
@@ -1,0 +1,2 @@
+from .config import ArborGRPOConfig  # noqa: F401
+from .trainer import ArborGRPOTrainer  # noqa: F401
