+++ arbor/training/sft/__init__.py
@@ -1,0 +1,1 @@
+from .trainer import *  # noqa: F403
