+++ arbor/training/utils/__init__.py
@@ -1,0 +1,1 @@
+from .arg_parser import get_training_arg_parser  # noqa: F401
