+++ arbor/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Shared utilities used across Arbor (server, client, training)."""
