+++ arbor/utils/format_detection.py
@@ -6,7 +6,7 @@ import json
 from pathlib import Path
 from typing import Any, Literal
 
-from arbor.server.utils.logging import get_logger
+from arbor.core.logging import get_logger
 
 logger = get_logger(__name__)
